Do not propagate profile changes of shadow-banned users into rooms. (#8157)
