+++ changelog.d/8157.feature
@@ -1,0 +1,1 @@
+Add support for shadow-banning users (ignoring any message send requests).
